Fixed warning on MSVC.
See issue #154

+++ multicore_engine_graphics_test/src/graphics/graphics_test.cpp
@@ -12,7 +12,7 @@ namespace graphics {
 
 graphics_test::graphics_test()
 		: glfw_win_("Vulkan Test", glm::ivec2(800, 600)), dev_(inst_), win_(inst_, glfw_win_, dev_),
-		  plc_(dev_), mem_mgr_(&dev_, 1 << 27), dqm_(&dev_, win_.swapchain_images().size()) {}
+		  plc_(dev_), mem_mgr_(&dev_, 1 << 27), dqm_(&dev_, uint32_t(win_.swapchain_images().size())) {}
 
 graphics_test::~graphics_test() {}
